pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipegate"
version = "0.1.0"
description = "A lightweight, self-hosted tunneling proxy — poor man's ngrok"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyjwt>=2.0",
    "orjson>=3.0",
    "async-timeout>=4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24,<1.0",
    "httpx>=0.27",
    "ruff>=0.9",
    "mypy>=1.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["async_timeout", "uvicorn"]
ignore_missing_imports = true
